requirements.txt
# Core dependencies for FPL Elo Insights
pandas>=2.3.0
numpy>=2.2.0
requests>=2.28.0
python-dotenv>=0.19.0

# Optional dependencies (uncomment if using Supabase)
# supabase>=1.0.0

# Development dependencies
jupyter>=1.0.0